requirements.txt
amqp==5.2.0
asgiref==3.6.0
async-timeout==4.0.3
beautifulsoup4==4.11.2
billiard==4.2.0
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crypto==1.4.1
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.5
django-allauth==0.52.0
django-bootstrap4==22.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
easycodefpy==1.0.1
idna==3.4
kombu==5.3.7
Naked==0.1.32
oauthlib==3.2.2
prompt_toolkit==3.0.47
protobuf==3.20.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.1
redis==5.0.6
requests==2.28.1
requests-oauthlib==1.3.1
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.3.0
gunicorn
python-dotenv
Pillow==9.4.0
ipython==8.7.0
drf_spectacular==0.28.0
pytest-django==4.11.1